+++ src/main/java/com/thinkgem/jeesite/modules/inventory/web/RoomController.java
@@ -232,8 +232,15 @@ public class RoomController extends BaseController {
 	@RequiresPermissions("inventory:room:edit")
 	@RequestMapping(value = "delete")
 	public String delete(Room room, RedirectAttributes redirectAttributes) {
-		roomService.delete(room);
-		addMessage(redirectAttributes, "删除房间及图片信息成功");
+		Room queryRoom = roomService.get(room);
+		String roomStatus = queryRoom.getRoomStatus();
+		// 2已预定 3已出租
+		if ("2".equals(roomStatus) || "3".equals(roomStatus)) {
+			addMessage(redirectAttributes, "房间已预定或已出租，不能删除！");
+		} else {
+			roomService.delete(room);
+			addMessage(redirectAttributes, "删除房间及图片信息成功");
+		}
 		return "redirect:" + Global.getAdminPath() + "/inventory/room/?repage";
 	}
